Fail on exception in nextHandler of succeeding and failing
VertxTestContext will immediately fail the test if nextHandler
of succeeding(nextHandler) or failing(nextHandler) throws an
exception (any Throwable).

Why should succeeding(nextHandler) and failing(nextHandler)
catch any exception that nextHandler throws?

We write tests to let VertxTestContext find and report bugs.
All failed assertions must fail the test.
If the test code or the code under test throws an exception
it is almost always unintended and a bug and should also fail
the test and report that exception including the stacktrace.

Exceptions should fail the test by default without the need
to manually wrap the test code into verify(...).

Why is this not a breaking change?

Currently it is not documented and therefore undefined how
VertxTestContext handles exceptions thrown by nextHandler:
https://github.com/vert-x3/vertx-junit5/blob/4.0.0-milestone4/vertx-junit5/src/main/java/io/vertx/junit5/VertxTestContext.java#L175
https://github.com/vert-x3/vertx-junit5/blob/4.0.0-milestone4/vertx-junit5/src/main/java/io/vertx/junit5/VertxTestContext.java#L206

How does this feature correspond to vertx-unit?

The implementations of vertx-junit TestContext
asyncAssertSuccess(Handler<T> resultHandler) and
asyncAssertFailure(Handler<Throwable> causeHandler) fail the
test if the handler throws an exception:
https://github.com/vert-x3/vertx-unit/blob/4.0.0-milestone4/src/main/java/io/vertx/ext/unit/impl/TestContextImpl.java#L318
https://github.com/vert-x3/vertx-unit/blob/4.0.0-milestone4/src/main/java/io/vertx/ext/unit/impl/TestContextImpl.java#L342

This (undocumented) vertx-unit feature should be carried forward
into vertx-junit5 succeeding and failing methods.

How does this feature improve test reporting ergonomics?

An exception that is ignored usually results in a
test timeout. The timeout message contains no indication
that there was an exception.

If the test fails because of an exception in the nextHandler
code block the exception message and the exception stack trace
is reported to the testing framework (for example JUnit).
This information is available on the command line and in IDEs
and help developers to pinpoint and fix the problem.

How does this feature improve test writing ergonomics?

It reduces the boilerplate.

```
testContext.succeeding(response -> testContext.verify(() -> {
  ...
}))
```

is reduced to

```
testContext.succeeding(response -> {
  ...
})
```

and therefore improves readability by avoiding to wrap the inner
test code into yet another lambda.

Developers cannot unintentionally forget to add vertify(...) when
using succeeding(...) or failing(...).

This feature is almost always the expected behaviour.

How can an exception still been ignored as before?

```
testContext.succeeding(response -> {
  try {
    // some test code
  } catch (Throwable e) {
    return;
  }
})
```

This pull request obsoletes #63