Remove succeeding() and failing(), add failingThenComplete()
Using the parameter-less method VertxTestContext#succeeding()
is almost always unintended and should be replaced by

* completing()
* succeeding(value -> checkpoint.flag())
* succeeding(value -> { more testing code })
* succeeding(value -> {})

The same for failing().

Removing the methods succeeding() and failing() improves
VertxTestContext's ergonomics because people are no longer
tempted to use them. The replacement either fixes an unintended
bug in people's unit test code or is more readable because
it explicitly states the intention.

Example for unintended succeeding() usage that is not executed:
https://github.com/reactiverse/reactiverse-junit5-extensions/pull/3/files

The new method failingThenComplete() is a short-cut for
failing(exception -> vertxTestContext.completeNow())
and parallels completing() that is an existing short-cut for
succeeding(value -> vertxTestContext.completeNow()).

Note that vertx-unit's TestContext#asyncAssertSuccess() and
TestContext#asyncAssertFailure() are different because they
complete the test context by design. People switching from
vertx-unit to vertx-junit5 will not fall into this trap if we
remove succeeding() and failing().

Fixes #77.

Signed-off-by: Julian Ladisch <eclipse.org-rtn@ladisch.de>